Upgrade procedures for Liberty to Newton
These are bascially the upgrade procedures that was used in our
OpenStack upgrade. This won't work by itself but it might give you
a good view of what was necessary.

The aggregate.py and flavor.py are the original Nova Newton files,
I will add a the patched ones in the next commit.